refactor: move i257 structure into the sub i257 mod in the primitives crate
