+++ crates/primitives/src/lib.rs
@@ -1,5 +1,4 @@
 use std::hash::{BuildHasher, BuildHasherDefault, Hash, Hasher};
-use std::ops::{Add, AddAssign, Sub, SubAssign};
 
 pub use alloy_evm::EvmEnv;
 pub use alloy_primitives::{
@@ -52,6 +51,52 @@ pub use revm::{Context, MainBuilder, MainContext, MainnetEvm};
 pub use revm::{ExecuteCommitEvm, ExecuteEvm, InspectEvm};
 pub use rustc_hash::FxBuildHasher;
 
+pub mod i257;
+pub use i257::I257;
+
+/// Converts a [U256] value to a [u64], saturating to [MAX][u64] if the value is too large.
+#[macro_export]
+macro_rules! as_u64_saturated {
+    ($v:expr) => {
+        match $v.as_limbs() {
+            x => {
+                if (x[1] == 0) & (x[2] == 0) & (x[3] == 0) {
+                    x[0]
+                } else {
+                    u64::MAX
+                }
+            }
+        }
+    };
+}
+
+/// Converts a [U256] value to a [usize], saturating to [MAX][usize] if the value is too large.
+#[macro_export]
+macro_rules! as_usize_saturated {
+    ($v:expr) => {
+        usize::try_from($crate::as_u64_saturated!($v)).unwrap_or(usize::MAX)
+    };
+}
+
+/// Converts a [U256] value to a [isize], saturating to [MAX][isize] if the value is too large.
+#[macro_export]
+macro_rules! as_isize_saturated {
+    ($v:expr) => {
+        isize::try_from($crate::as_u64_saturated!($v)).unwrap_or(isize::MAX)
+    };
+}
+
+/// `const` Option `?`.
+#[macro_export]
+macro_rules! tri {
+    ($e:expr) => {
+        match $e {
+            Some(v) => v,
+            None => return None,
+        }
+    };
+}
+
 #[derive(Clone, Debug, Default, PartialEq, Eq)]
 pub struct Env {
     pub block: BlockEnv,
@@ -87,22 +132,6 @@ impl Hasher for SuffixHasher {
 /// Build a suffix hasher
 pub type BuildSuffixHasher = BuildHasherDefault<SuffixHasher>;
 
-/// Converts a [U256] value to a [u64], saturating to [MAX][u64] if the value is too large.
-#[macro_export]
-macro_rules! as_u64_saturated {
-    ($v:expr) => {
-        match $v.as_limbs() {
-            x => {
-                if (x[1] == 0) & (x[2] == 0) & (x[3] == 0) {
-                    x[0]
-                } else {
-                    u64::MAX
-                }
-            }
-        }
-    };
-}
-
 /// This is primarily used for memory location hash, but can also be used for
 /// transaction indexes, etc.
 #[derive(Debug, Default)]
@@ -130,135 +159,4 @@ pub type BuildIdentityHasher = BuildHasherDefault<IdentityHasher>;
 #[inline(always)]
 pub fn hash_deterministic<T: Hash>(x: T) -> u64 {
     FxBuildHasher.hash_one(x)
-}
-
-/// Converts a [U256] value to a [usize], saturating to [MAX][usize] if the value is too large.
-#[macro_export]
-macro_rules! as_usize_saturated {
-    ($v:expr) => {
-        usize::try_from($crate::as_u64_saturated!($v)).unwrap_or(usize::MAX)
-    };
-}
-
-/// Converts a [U256] value to a [isize], saturating to [MAX][isize] if the value is too large.
-#[macro_export]
-macro_rules! as_isize_saturated {
-    ($v:expr) => {
-        isize::try_from($crate::as_u64_saturated!($v)).unwrap_or(isize::MAX)
-    };
-}
-
-/// `const` Option `?`.
-#[macro_export]
-macro_rules! tri {
-    ($e:expr) => {
-        match $e {
-            Some(v) => v,
-            None => return None,
-        }
-    };
-}
-
-/// A 257-bit signed integer type.
-/// It is represented as a 256-bit unsigned integer and a boolean
-/// indicating whether the number is negative.
-#[derive(Clone, Debug, PartialEq, Eq, Copy)]
-pub struct I257 {
-    value: U256,
-    is_negative: bool,
-}
-
-impl I257 {
-    #[inline]
-    pub fn new(value: U256, is_negative: bool) -> Self {
-        I257 { value, is_negative }
-    }
-
-    #[inline]
-    pub fn abs_value(&self) -> U256 {
-        self.value
-    }
-
-    #[inline]
-    pub fn sign(&self) -> i8 {
-        if self.is_negative { -1 } else { 1 }
-    }
-
-    pub const ZERO: I257 = I257 {
-        value: U256::ZERO,
-        is_negative: false,
-    };
-
-    pub const NEGATIVE_ZERO: I257 = I257 {
-        value: U256::ZERO,
-        is_negative: true,
-    };
-}
-
-impl From<U256> for I257 {
-    fn from(val: U256) -> Self {
-        I257 {
-            value: val,
-            is_negative: false,
-        }
-    }
-}
-
-/// Implementing the `Add` and `Sub` traits for I257 to handle addition and subtraction
-impl Add for I257 {
-    type Output = I257;
-
-    fn add(self, other: I257) -> I257 {
-        if self.is_negative == other.is_negative {
-            I257 {
-                value: self.value.saturating_add(other.value),
-                is_negative: self.is_negative,
-            }
-        } else if self.abs_value() > other.abs_value() {
-            I257 {
-                value: self.value - other.value,
-                is_negative: self.is_negative,
-            }
-        } else {
-            I257 {
-                value: other.value - self.value,
-                is_negative: other.is_negative,
-            }
-        }
-    }
-}
-
-impl AddAssign for I257 {
-    fn add_assign(&mut self, other: I257) {
-        *self = *self + other;
-    }
-}
-
-impl Sub for I257 {
-    type Output = I257;
-
-    fn sub(self, other: I257) -> I257 {
-        if self.is_negative != other.is_negative {
-            I257 {
-                value: self.value.saturating_add(other.value),
-                is_negative: self.is_negative,
-            }
-        } else if self.abs_value() > other.abs_value() {
-            I257 {
-                value: self.value - other.value,
-                is_negative: self.is_negative,
-            }
-        } else {
-            I257 {
-                value: other.value - self.value,
-                is_negative: !self.is_negative,
-            }
-        }
-    }
-}
-
-impl SubAssign for I257 {
-    fn sub_assign(&mut self, other: I257) {
-        *self = *self - other;
-    }
 }
